// src/core/types.ts
export interface DuoRect {
  readonly x: number;
  readonly y: number;
  readonly width: number;
  readonly height: number;
}